[Decomp][Player] Cohesion split: extract storage system to PlayerItem.cpp
