+++ src/game/Object/Player.cpp
@@ -6926,411 +6926,16 @@ void Player::SendPetSkillWipeConfirm()
  * @param item The item to reflect in the virtual slot.
  */
 
-/**
- * @brief Checks whether a bag and slot refer to normal inventory storage.
- *
- * @param bag The bag identifier.
- * @param slot The slot identifier.
- * @return True if the position is an inventory position; otherwise, false.
- */
-bool Player::IsInventoryPos(uint8 bag, uint8 slot)
-{
-    if (bag == INVENTORY_SLOT_BAG_0 && slot == NULL_SLOT)
-    {
-        return true;
-    }
-    if (bag == INVENTORY_SLOT_BAG_0 && (slot >= INVENTORY_SLOT_ITEM_START && slot < INVENTORY_SLOT_ITEM_END))
-    {
-        return true;
-    }
-    if (bag >= INVENTORY_SLOT_BAG_START && bag < INVENTORY_SLOT_BAG_END)
-    {
-        return true;
-    }
-    if (bag == INVENTORY_SLOT_BAG_0 && (slot >= KEYRING_SLOT_START && slot < CURRENCYTOKEN_SLOT_END))
-    {
-        return true;
-    }
-    return false;
-}
-
-/**
- * @brief Checks whether a bag and slot refer to an equipment position.
- *
- * @param bag The bag identifier.
- * @param slot The slot identifier.
- * @return True if the position is an equipment position; otherwise, false.
- */
-bool Player::IsEquipmentPos(uint8 bag, uint8 slot)
-{
-    if (bag == INVENTORY_SLOT_BAG_0 && (slot < EQUIPMENT_SLOT_END))
-    {
-        return true;
-    }
-    if (bag == INVENTORY_SLOT_BAG_0 && (slot >= INVENTORY_SLOT_BAG_START && slot < INVENTORY_SLOT_BAG_END))
-    {
-        return true;
-    }
-    return false;
-}
-
-/**
- * @brief Checks whether a bag and slot refer to bank storage.
- *
- * @param bag The bag identifier.
- * @param slot The slot identifier.
- * @return True if the position is in bank storage; otherwise, false.
- */
-bool Player::IsBankPos(uint8 bag, uint8 slot)
-{
-    if (bag == INVENTORY_SLOT_BAG_0 && (slot >= BANK_SLOT_ITEM_START && slot < BANK_SLOT_ITEM_END))
-    {
-        return true;
-    }
-    if (bag == INVENTORY_SLOT_BAG_0 && (slot >= BANK_SLOT_BAG_START && slot < BANK_SLOT_BAG_END))
-    {
-        return true;
-    }
-    if (bag >= BANK_SLOT_BAG_START && bag < BANK_SLOT_BAG_END)
-    {
-        return true;
-    }
-    return false;
-}
-
-/**
- * @brief Checks whether a packed position refers to a bag slot.
- *
- * @param pos The packed bag and slot position.
- * @return True if the position is a bag slot; otherwise, false.
- */
-bool Player::IsBagPos(uint16 pos)
-{
-    uint8 bag = pos >> 8;
-    uint8 slot = pos & 255;
-    if (bag == INVENTORY_SLOT_BAG_0 && (slot >= INVENTORY_SLOT_BAG_START && slot < INVENTORY_SLOT_BAG_END))
-    {
-        return true;
-    }
-    if (bag == INVENTORY_SLOT_BAG_0 && (slot >= BANK_SLOT_BAG_START && slot < BANK_SLOT_BAG_END))
-    {
-        return true;
-    }
-    return false;
-}
-
-/**
- * @brief Checks whether a bag and slot form a valid player storage position.
- *
- * @param bag The bag identifier.
- * @param slot The slot identifier.
- * @param explicit_pos True if the caller requires an explicit fixed position.
- * @return True if the position is valid; otherwise, false.
- */
-bool Player::IsValidPos(uint8 bag, uint8 slot, bool explicit_pos) const
-{
-    // post selected
-    if (bag == NULL_BAG && !explicit_pos)
-    {
-        return true;
-    }
-
-    if (bag == INVENTORY_SLOT_BAG_0)
-    {
-        // any post selected
-        if (slot == NULL_SLOT && !explicit_pos)
-        {
-            return true;
-        }
-
-        // equipment
-        if (slot < EQUIPMENT_SLOT_END)
-        {
-            return true;
-        }
-
-        // bag equip slots
-        if (slot >= INVENTORY_SLOT_BAG_START && slot < INVENTORY_SLOT_BAG_END)
-        {
-            return true;
-        }
-
-        // backpack slots
-        if (slot >= INVENTORY_SLOT_ITEM_START && slot < INVENTORY_SLOT_ITEM_END)
-        {
-            return true;
-        }
-
-        // keyring slots
-        if (slot >= KEYRING_SLOT_START && slot < KEYRING_SLOT_END)
-        {
-            return true;
-        }
-
-        // bank main slots
-        if (slot >= BANK_SLOT_ITEM_START && slot < BANK_SLOT_ITEM_END)
-        {
-            return true;
-        }
-
-        // bank bag slots
-        if (slot >= BANK_SLOT_BAG_START && slot < BANK_SLOT_BAG_END)
-        {
-            return true;
-        }
-
-        return false;
-    }
-
-    // bag content slots
-    if (bag >= INVENTORY_SLOT_BAG_START && bag < INVENTORY_SLOT_BAG_END)
-    {
-        Bag* pBag = reinterpret_cast<Bag*>(GetItemByPos(INVENTORY_SLOT_BAG_0, bag));
-        if (!pBag)
-        {
-            return false;
-        }
-
-        // any post selected
-        if (slot == NULL_SLOT && !explicit_pos)
-        {
-            return true;
-        }
-
-        return slot < pBag->GetBagSize();
-    }
-
-    // bank bag content slots
-    if (bag >= BANK_SLOT_BAG_START && bag < BANK_SLOT_BAG_END)
-    {
-        Bag* pBag = reinterpret_cast<Bag*>(GetItemByPos(INVENTORY_SLOT_BAG_0, bag));
-        if (!pBag)
-        {
-            return false;
-        }
-
-        // any post selected
-        if (slot == NULL_SLOT && !explicit_pos)
-        {
-            return true;
-        }
-
-        return slot < pBag->GetBagSize();
-    }
-
-    // where this?
-    return false;
-}
-
-/**
- * @brief Checks whether the player owns at least a given count of an item.
- *
- * @param item The item entry to count.
- * @param count The required quantity.
- * @param inBankAlso True to include bank storage in the search.
- * @return True if enough items are present; otherwise, false.
- */
-bool Player::HasItemCount(uint32 item, uint32 count, bool inBankAlso) const
-{
-    uint32 tempcount = 0;
-    for (int i = EQUIPMENT_SLOT_START; i < INVENTORY_SLOT_ITEM_END; ++i)
-    {
-        Item* pItem = GetItemByPos(INVENTORY_SLOT_BAG_0, i);
-        if (pItem && pItem->GetEntry() == item && !pItem->IsInTrade())
-        {
-            tempcount += pItem->GetCount();
-            if (tempcount >= count)
-            {
-                return true;
-            }
-        }
-    }
-    for (int i = KEYRING_SLOT_START; i < CURRENCYTOKEN_SLOT_END; ++i)
-    {
-        Item* pItem = GetItemByPos(INVENTORY_SLOT_BAG_0, i);
-        if (pItem && pItem->GetEntry() == item && !pItem->IsInTrade())
-        {
-            tempcount += pItem->GetCount();
-            if (tempcount >= count)
-            {
-                return true;
-            }
-        }
-    }
-    for (int i = INVENTORY_SLOT_BAG_START; i < INVENTORY_SLOT_BAG_END; ++i)
-    {
-        if (Bag* pBag = reinterpret_cast<Bag*>(GetItemByPos(INVENTORY_SLOT_BAG_0, i)))
-        {
-            for (uint32 j = 0; j < pBag->GetBagSize(); ++j)
-            {
-                Item* pItem = GetItemByPos(i, j);
-                if (pItem && pItem->GetEntry() == item && !pItem->IsInTrade())
-                {
-                    tempcount += pItem->GetCount();
-                    if (tempcount >= count)
-                    {
-                        return true;
-                    }
-                }
-            }
-        }
-    }
-
-    if (inBankAlso)
-    {
-        for (int i = BANK_SLOT_ITEM_START; i < BANK_SLOT_ITEM_END; ++i)
-        {
-            Item* pItem = GetItemByPos(INVENTORY_SLOT_BAG_0, i);
-            if (pItem && pItem->GetEntry() == item && !pItem->IsInTrade())
-            {
-                tempcount += pItem->GetCount();
-                if (tempcount >= count)
-                {
-                    return true;
-                }
-            }
-        }
-        for (int i = BANK_SLOT_BAG_START; i < BANK_SLOT_BAG_END; ++i)
-        {
-            if (Bag* pBag = reinterpret_cast<Bag*>(GetItemByPos(INVENTORY_SLOT_BAG_0, i)))
-            {
-                for (uint32 j = 0; j < pBag->GetBagSize(); ++j)
-                {
-                    Item* pItem = GetItemByPos(i, j);
-                    if (pItem && pItem->GetEntry() == item && !pItem->IsInTrade())
-                    {
-                        tempcount += pItem->GetCount();
-                        if (tempcount >= count)
-                        {
-                            return true;
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    return false;
-}
-
-bool Player::HasItemOrGemWithIdEquipped(uint32 item, uint32 count, uint8 except_slot) const
-{
-    uint32 tempcount = 0;
-    for (int i = EQUIPMENT_SLOT_START; i < EQUIPMENT_SLOT_END; ++i)
-    {
-        if (i == int(except_slot))
-        {
-            continue;
-        }
-
-        Item* pItem = GetItemByPos(INVENTORY_SLOT_BAG_0, i);
-        if (pItem && pItem->GetEntry() == item)
-        {
-            tempcount += pItem->GetCount();
-            if (tempcount >= count)
-            {
-                return true;
-            }
-        }
-    }
-
-    ItemPrototype const* pProto = ObjectMgr::GetItemPrototype(item);
-    if (pProto && pProto->GemProperties)
-    {
-        for (int i = EQUIPMENT_SLOT_START; i < EQUIPMENT_SLOT_END; ++i)
-        {
-            if (i == int(except_slot))
-            {
-                continue;
-            }
-
-            Item* pItem = GetItemByPos(INVENTORY_SLOT_BAG_0, i);
-            if (pItem && pItem->GetProto()->Socket[0].Color)
-            {
-                tempcount += pItem->GetGemCountWithID(item);
-                if (tempcount >= count)
-                {
-                    return true;
-                }
-            }
-        }
-    }
-
-    return false;
-}
-
-bool Player::HasItemOrGemWithLimitCategoryEquipped(uint32 limitCategory, uint32 count, uint8 except_slot) const
-{
-    uint32 tempcount = 0;
-    for (int i = EQUIPMENT_SLOT_START; i < EQUIPMENT_SLOT_END; ++i)
-    {
-        if (i == int(except_slot))
-        {
-            continue;
-        }
-
-        Item* pItem = GetItemByPos(INVENTORY_SLOT_BAG_0, i);
-        if (!pItem)
-        {
-            continue;
-        }
-
-        ItemPrototype const* pProto = pItem->GetProto();
-        if (!pProto)
-        {
-            continue;
-        }
-
-        if (pProto->ItemLimitCategory == limitCategory)
-        {
-            tempcount += pItem->GetCount();
-            if (tempcount >= count)
-            {
-                return true;
-            }
-        }
-
-        if (pProto->Socket[0].Color)
-        {
-            tempcount += pItem->GetGemCountWithLimitCategory(limitCategory);
-            if (tempcount >= count)
-            {
-                return true;
-            }
-        }
-    }
-
-    return false;
-}
 
 
 
-/**
- * @brief Cancels the current trade and clears trade state for both players.
- *
- * @param sendback True to send a cancel notification to this player as well.
- */
-void Player::TradeCancel(bool sendback)
-{
-    if (m_trade)
-    {
-        Player* trader = m_trade->GetTrader();
 
-        // send yellow "Trade canceled" message to both traders
-        if (sendback)
-        {
-            GetSession()->SendCancelTrade();
-        }
 
-        trader->GetSession()->SendCancelTrade();
 
-        // cleanup
-        delete m_trade;
-        m_trade = NULL;
-        delete trader->m_trade;
-        trader->m_trade = NULL;
-    }
-}
+
+
+
+
 
 
 /*********************************************************/
